Trying out a new concept, called WindowedSearch.
With it, we can associate metadata for our clients so they could find the link to the next and previous resource.
This might be not necessary as based on query conditions the client could easily deduce the URLs that they care about anyway. Interestingly, we're able to tell the client whether there will be any more data available in their search. Maybe that is exactly what we should say in a header. It is quite nice to have that included for client-side. Though it might be even nicer to give the client a result count, e.g. how many previous games there will be. And how many next games there will be.

Also, need to think about whether we're interested in the 'after' lookup at all. It's not actually that useful right now for the site. Could simply remove it together with 'first' and stick to 'recent' and 'before', since they are all that we really technically need to fulfil the needs.